requirements.txt
PySide6
google-generativeai
docxtpl
python-docx
docx2pdf